Move Sqlite code into Schema/Dialect
