+++ lib/Cake/Test/TestCase/Database/Driver/SqliteTest.php
@@ -29,16 +29,6 @@ use \PDO;
 class SqliteTest extends TestCase {
 
 /**
- * Helper method for skipping tests that need a real connection.
- *
- * @return void
- */
-	protected function _needsConnection() {
-		$config = Configure::read('Datasource.test');
-		$this->skipIf(strpos($config['datasource'], 'Sqlite') === false, 'Not using Sqlite for test config');
-	}
-
-/**
  * Test connecting to Sqlite with default configuration
  *
  * @return void
@@ -101,188 +91,5 @@ class SqliteTest extends TestCase {
 		$driver->connect($config);
 	}
 
-/**
- * Dataprovider for column testing
- *
- * @return array
- */
-	public static function columnProvider() {
-		return [
-			[
-				'DATETIME',
-				['datetime', null]
-			],
-			[
-				'DATE',
-				['date', null]
-			],
-			[
-				'TIME',
-				['time', null]
-			],
-			[
-				'BOOLEAN',
-				['boolean', null]
-			],
-			[
-				'BIGINT',
-				['biginteger', null]
-			],
-			[
-				'VARCHAR(255)',
-				['string', 255]
-			],
-			[
-				'CHAR(25)',
-				['string', 25]
-			],
-			[
-				'BLOB',
-				['binary', null]
-			],
-			[
-				'INTEGER(11)',
-				['integer', 11]
-			],
-			[
-				'TINYINT(5)',
-				['integer', 5]
-			],
-			[
-				'MEDIUMINT(10)',
-				['integer', 10]
-			],
-			[
-				'FLOAT',
-				['float', null]
-			],
-			[
-				'DOUBLE',
-				['float', null]
-			],
-			[
-				'REAL',
-				['float', null]
-			],
-			[
-				'DECIMAL(11,2)',
-				['decimal', null]
-			],
-		];
-	}
-
-/**
- * Test parsing SQLite column types.
- *
- * @dataProvider columnProvider
- * @return void
- */
-	public function testConvertColumnType($input, $expected) {
-		$driver = new Sqlite();
-		$this->assertEquals($expected, $driver->convertColumn($input));
-	}
-
-/**
- * Creates tables for testing listTables/describe()
- *
- * @param Connection $connection
- * @return void
- */
-	protected function _createTables($connection) {
-		$this->_needsConnection();
-		$connection->execute('DROP TABLE IF EXISTS articles');
-		$connection->execute('DROP TABLE IF EXISTS authors');
-
-		$table = <<<SQL
-CREATE TABLE authors(
-id INTEGER PRIMARY KEY AUTOINCREMENT,
-name VARCHAR(50),
-bio TEXT,
-created DATETIME
-)
-SQL;
-		$connection->execute($table);
-
-		$table = <<<SQL
-CREATE TABLE articles(
-id INTEGER PRIMARY KEY AUTOINCREMENT,
-title VARCHAR(20) DEFAULT 'testing',
-body TEXT,
-author_id INT(11) NOT NULL,
-published BOOLEAN DEFAULT 0,
-created DATETIME
-)
-SQL;
-		$connection->execute($table);
-	}
-
-/**
- * Test listing tables with Sqlite
- *
- * @return void
- */
-	public function testListTables() {
-		$connection = new Connection(Configure::read('Datasource.test'));
-		$this->_createTables($connection);
-
-		$result = $connection->listTables();
-		$this->assertInternalType('array', $result);
-		$this->assertCount(3, $result);
-		$this->assertEquals('articles', $result[0]);
-		$this->assertEquals('authors', $result[1]);
-		$this->assertEquals('sqlite_sequence', $result[2]);
-	}
-
-/**
- * Test describing a table with Sqlite
- *
- * @return void
- */
-	public function testDescribeTable() {
-		$connection = new Connection(Configure::read('Datasource.test'));
-		$this->_createTables($connection);
-
-		$result = $connection->describe('articles');
-		$expected = [
-			'id' => [
-				'type' => 'integer',
-				'null' => false,
-				'default' => null,
-				'length' => null,
-				'key' => 'primary',
-			],
-			'title' => [
-				'type' => 'string',
-				'null' => true,
-				'default' => 'testing',
-				'length' => 20,
-			],
-			'body' => [
-				'type' => 'text',
-				'null' => true,
-				'default' => null,
-				'length' => null,
-			],
-			'author_id' => [
-				'type' => 'integer',
-				'null' => false,
-				'default' => null,
-				'length' => 11,
-			],
-			'published' => [
-				'type' => 'boolean',
-				'null' => true,
-				'default' => 0,
-				'length' => null,
-			],
-			'created' => [
-				'type' => 'datetime',
-				'null' => true,
-				'default' => null,
-				'length' => null,
-			],
-		];
-		$this->assertEquals($expected, $result);
-	}
 
 }
